+++ src/main/java/chess/controller/api/GameController.java
@@ -1,13 +1,17 @@
 package chess.controller.api;
 
 import chess.dto.ChessGameDto;
+import chess.dto.ChessGamesDto;
 import chess.dto.MoveDto;
+import chess.dto.TitleDto;
 import chess.service.GameService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
+import java.util.stream.Collectors;
+
 @RestController
-@RequestMapping("/game")
 public class GameController {
 
     private final GameService gameService;
@@ -16,25 +20,41 @@ public class GameController {
         this.gameService = gameService;
     }
 
-    @GetMapping("/{id}/game-info")
+    @GetMapping("/game/{id}/game-info")
     public ResponseEntity<ChessGameDto> loadGame(@PathVariable Long id) {
         return ResponseEntity.ok(new ChessGameDto(gameService.loadGame(id)));
     }
 
-    @PutMapping(path = "/{id}/move")
+    @PutMapping(path = "/game/{id}/move")
     public ResponseEntity<ChessGameDto> move(@PathVariable Long id, @RequestBody MoveDto moveDto) {
         gameService.move(id, moveDto);
         return loadGame(id);
     }
 
-    @PostMapping("/{id}/terminate")
+    @PostMapping("/game/{id}/terminate")
     public ResponseEntity<Void> terminateGame(@PathVariable Long id) {
         gameService.terminateGame(id);
         return ResponseEntity.ok().build();
     }
 
-    @PostMapping("/{id}/restart")
+    @PostMapping("/game/{id}/restart")
     public ResponseEntity<ChessGameDto> restart(@PathVariable Long id) {
         return ResponseEntity.ok(new ChessGameDto(gameService.restart(id)));
+    }
+
+    @PostMapping("/game")
+    public ResponseEntity<Long> newGame(@RequestBody @Valid TitleDto titleDto) {
+        gameService.verifyDuplicateTitleInGames(titleDto.getTitle());
+        return ResponseEntity.ok(gameService.newGame(titleDto.getTitle()));
+    }
+
+    @GetMapping("/games")
+    public ResponseEntity<ChessGamesDto> findAllRooms() {
+        return ResponseEntity.ok(new ChessGamesDto(
+                gameService.findAllGames()
+                        .stream()
+                        .map(ChessGameDto::new)
+                        .collect(Collectors.toList())
+        ));
     }
 }
